Fixes from Andrew Galante for build with gcc4. LWOB materials finished, some research for LWO2. Unit tests are working again. Added makefile and project workspace for xcode3 (from Andrew, too).
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@110 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ include/aiMaterial.inl
@@ -72,7 +72,7 @@ inline aiReturn aiMaterial::Get(const char* pKey,Type* pOut,
 		if (prop->mDataLength < sizeof(Type)*iNum)return AI_FAILURE;
 		if (strcmp(prop->mData,(char*)aiPTI_Buffer)!=0)return AI_FAILURE;
 
-		iNum = std::min(iNum,prop->mDataLength / sizeof(Type));
+		iNum = std::min((size_t)iNum,prop->mDataLength / sizeof(Type));
 		::memcpy(pOut,prop->mData,iNum * sizeof(Type));
 		if (pMax)*pMax = iNum;
 	}
